Expand graphics api

+++ Engine/AABB.cpp
@@ -59,9 +59,9 @@ Region AABB::CalcScreenSpaceCoordinates(const AABB& aabb, const glm::mat4& model
     min = glm::clamp(min, glm::vec2(), glm::vec2(viewport[2] - 1, viewport[3] - 1));
     max = glm::clamp(max, glm::vec2(), glm::vec2(viewport[2] - 1, viewport[3] - 1));
 
-    size_t x = static_cast<size_t>(std::min(min.x, max.x));
-    size_t y = static_cast<size_t>(std::min(min.y, max.y));
-    size_t w = static_cast<size_t>(std::abs(max.x - min.x));
-    size_t h = static_cast<size_t>(std::abs(max.y - min.y));
+    unsigned x = static_cast<unsigned>(std::min(min.x, max.x));
+    unsigned y = static_cast<unsigned>(std::min(min.y, max.y));
+    unsigned w = static_cast<unsigned>(std::abs(max.x - min.x));
+    unsigned h = static_cast<unsigned>(std::abs(max.y - min.y));
     return { x, y, w, h };
 }
